Make walk() method of _AD_group return non-user members
invoking walk() on an _AD_group object would work with user groups, but for computer groups it would not list members. The walk method was explicitly checking for a class of 'user' to find the non-subgroup members of a group. To fix it, I changed it to check that the class is not 'group'.